Added SessionID to URL link, Added presenter UI screen
