+++ Hungry-Hippo-Game/src/pages/PresenterPage/Presenter.module.css
@@ -1,0 +1,46 @@
+.containerImg {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center; /* centers vertically */
+  height: 100vh; /* full screen height */
+  background-image: url('/assets/mainBackground.png');
+  background-position: center;
+  background-repeat: no-repeat;
+  background-size: cover; 
+}
+
+.sessionText {
+  color: black;
+  margin-bottom: 1rem;
+}
+
+.sessionText2 {
+  color: black;
+  margin-bottom: 1rem;
+}
+
+.backButton {
+  background: none;
+  border: none;
+  cursor: pointer;
+  margin-bottom: 1rem;
+  position: absolute;
+  top: 20px;
+  left: 20px;
+  color: #333; /* adjust as needed */
+}
+
+.backButton:hover {
+  color: #555;
+}
+
+.backContent {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+}
+
+.backText {
+  font-size: 20px;
+}
